AI_3D_FailDetect_Raspberry_HA/requirements.txt
ultralytics==8.3.75
opencv-python
PyYAML
Flask
requests